--- project/src/components/error-message/error-message.tsx
@@ -1,9 +1,0 @@
-import {useAppSelector} from '../../store';
-
-export default function ErrorMessage(): JSX.Element | null {
-  const error = useAppSelector((state) => state.error);
-  return (error)
-    ? <div className='error-message'>{error}</div>
-    : null;
-
-}
