+++ project/src/pages/layout/layout.tsx
@@ -1,0 +1,11 @@
+import Header from '../../components/header/header';
+import {Main} from '../main-page/main';
+
+export default function Layout(): JSX.Element {
+  return (
+    <>
+      <Header />
+      <Main />
+    </>
+  );
+}
