+++ project/src/pages/room-page/room.tsx
@@ -1,6 +1,5 @@
 import { useParams} from 'react-router-dom';
 import { AuthorizationStatus, PROPERTY_RATING} from '../../const';
-import {Form} from '../../components/form/form';
 import {ReviewsList} from '../../components/reviews-list/reviews-list';
 import Map from '../../components/map/map';
 import {OffersList} from '../../components/offers-list/offers-list';
@@ -9,22 +8,22 @@ import {useEffect} from 'react';
 import { OfferType} from '../../types/offer-type';
 import {ReviewType} from '../../types/reviews';
 import {Loading} from '../../components/loading/loading';
-import Header from '../../components/header/header';
 import {fetchCurrentOfferAction, fetchNearOffersAction} from '../../services/api-actions';
 import {useDispatch} from 'react-redux';
+import Form from '../../components/form/form';
 
 export function Room(): JSX.Element {
   const {id} = useParams();
   const offerId = Number(id);
   const dispatch = useDispatch();
 
-  const isAuth: boolean = useAppSelector((state) => state.authorizationStatus) === AuthorizationStatus.Auth;
+  const isAuth: boolean = useAppSelector((state) => state.user.authorizationStatus) === AuthorizationStatus.Auth;
 
-  const currentOffer = useAppSelector((state) => state.offer) as OfferType;
+  const currentOffer = useAppSelector((state) => state.offer.offer) as OfferType;
 
-  const nearOffers = useAppSelector((state) => state.nearOffers);
+  const nearOffers = useAppSelector((state) => state.offers.nearOffers);
 
-  const reviews: ReviewType[] = useAppSelector((state) => state.reviews);
+  const reviews: ReviewType[] = useAppSelector((state) => state.offer.reviews);
 
   const root = document.getElementById('root') as HTMLElement;
   root.style.cssText = '';
@@ -36,108 +35,105 @@ export function Room(): JSX.Element {
 
   if(currentOffer) {
     return (
-      <>
-        <Header />
-        <main className="page__main page__main--property">
-          <section className="property">
-            <div className="property__gallery-container container">
-              <div className="property__gallery">
-                {currentOffer.images.slice(0, 6).map((img) => (
-                  <div key={img} className="property__image-wrapper">
-                    <img className="property__image" src={img} alt="studio"/>
-                  </div>
-                ))}
-              </div>
+      <main className="page__main page__main--property">
+        <section className="property">
+          <div className="property__gallery-container container">
+            <div className="property__gallery">
+              {currentOffer.images.slice(0, 6).map((img) => (
+                <div key={img} className="property__image-wrapper">
+                  <img className="property__image" src={img} alt="studio"/>
+                </div>
+              ))}
             </div>
-            <div className="property__container container">
-              <div className="property__wrapper">
-                {currentOffer?.isPremium && (
-                  <div className="property__mark">
-                    <span>Premium</span>
-                  </div>
-                )}
-                <div className="property__name-wrapper">
-                  <h1 className="property__name">
-                    {currentOffer?.name}
-                  </h1>
-                </div>
-                <div className="property__rating rating">
-                  <div className="property__stars rating__stars">
-                    <span style={{width: `${currentOffer.rating * PROPERTY_RATING}%`}}></span>
-                    <span className="visually-hidden">Rating</span>
-                  </div>
-                  <span className="property__rating-value rating__value">{currentOffer?.rating}</span>
-                </div>
-                <ul className="property__features">
-                  <li className="property__feature property__feature--entire">
-                    {currentOffer?.type}
-                  </li>
-                  <li className="property__feature property__feature--bedrooms">
-                    {currentOffer?.bedrooms} Bedrooms
-                  </li>
-                  <li className="property__feature property__feature--adults">
-                Max {currentOffer?.maxAdults} adults
-                  </li>
-                </ul>
-                <div className="property__price">
-                  <b className="property__price-value">{currentOffer?.price}</b>
-                  <span className="property__price-text">&nbsp;night</span>
-                </div>
-                <div className="property__inside">
-                  <h2 className="property__inside-title">What&apos;s inside</h2>
-                  <ul className="property__inside-list">
-                    {currentOffer?.goods.map((good) => (
-                      <li key={good} className="property__inside-item">
-                        {good}
-                      </li>
-                    ))}
-                  </ul>
-                </div>
-                <div className="property__host">
-                  <h2 className="property__host-title">Meet the host</h2>
-                  <div className="property__host-user user">
-                    <div className="property__avatar-wrapper property__avatar-wrapper--pro user__avatar-wrapper">
-                      <img className="property__avatar user__avatar" src={currentOffer?.host.avatarUrl} width="74"
-                        height="74"
-                        alt="Host avatar"
-                      />
-                    </div>
-                    <span className="property__user-name">
-                      {currentOffer?.host.name}
-                    </span>
-                    {currentOffer?.host.isPro && (
-                      <span className="property__user-status">
-                      Pro
-                      </span>
-                    )}
-                  </div>
-                  <div className="property__description">
-                    <p className="property__text">
-                      {currentOffer?.description}
-                    </p>
-                  </div>
-                </div>
-                <ReviewsList reviews={reviews}/>
-                {isAuth ? (
-                  <Form hotelId={currentOffer.id}/>
-                ) : (
-                  <div className="reviews__form">
-                  </div>
-                )}
-              </div>
-            </div>
-            <section className="property__map map">
-              <Map city={currentOffer.city.location} points={[...nearOffers, currentOffer]} selectedPoint={currentOffer}/>
-            </section>
-          </section>
-          <div className="container">
-            <section className="near-places places">
-              <h2 className="near-places__title">Other places in the neighbourhood</h2>
-              <OffersList offers={[...nearOffers]} />
-            </section>
           </div>
-        </main>
-      </>
+          <div className="property__container container">
+            <div className="property__wrapper">
+              {currentOffer?.isPremium && (
+                <div className="property__mark">
+                  <span>Premium</span>
+                </div>
+              )}
+              <div className="property__name-wrapper">
+                <h1 className="property__name">
+                  {currentOffer?.name}
+                </h1>
+              </div>
+              <div className="property__rating rating">
+                <div className="property__stars rating__stars">
+                  <span style={{width: `${currentOffer.rating * PROPERTY_RATING}%`}}></span>
+                  <span className="visually-hidden">Rating</span>
+                </div>
+                <span className="property__rating-value rating__value">{currentOffer?.rating}</span>
+              </div>
+              <ul className="property__features">
+                <li className="property__feature property__feature--entire">
+                  {currentOffer?.type}
+                </li>
+                <li className="property__feature property__feature--bedrooms">
+                  {currentOffer?.bedrooms} Bedrooms
+                </li>
+                <li className="property__feature property__feature--adults">
+                Max {currentOffer?.maxAdults} adults
+                </li>
+              </ul>
+              <div className="property__price">
+                <b className="property__price-value">{currentOffer?.price}</b>
+                <span className="property__price-text">&nbsp;night</span>
+              </div>
+              <div className="property__inside">
+                <h2 className="property__inside-title">What&apos;s inside</h2>
+                <ul className="property__inside-list">
+                  {currentOffer?.goods.map((good) => (
+                    <li key={good} className="property__inside-item">
+                      {good}
+                    </li>
+                  ))}
+                </ul>
+              </div>
+              <div className="property__host">
+                <h2 className="property__host-title">Meet the host</h2>
+                <div className="property__host-user user">
+                  <div className="property__avatar-wrapper property__avatar-wrapper--pro user__avatar-wrapper">
+                    <img className="property__avatar user__avatar" src={currentOffer?.host.avatarUrl} width="74"
+                      height="74"
+                      alt="Host avatar"
+                    />
+                  </div>
+                  <span className="property__user-name">
+                    {currentOffer?.host.name}
+                  </span>
+                  {currentOffer?.host.isPro && (
+                    <span className="property__user-status">
+                      Pro
+                    </span>
+                  )}
+                </div>
+                <div className="property__description">
+                  <p className="property__text">
+                    {currentOffer?.description}
+                  </p>
+                </div>
+              </div>
+              <ReviewsList reviews={reviews}/>
+              {isAuth ? (
+                <Form hotelId={currentOffer.id}/>
+              ) : (
+                <div className="reviews__form">
+                </div>
+              )}
+            </div>
+          </div>
+          <section className="property__map map">
+            <Map city={currentOffer.city.location} points={[...nearOffers, currentOffer]} selectedPoint={currentOffer}/>
+          </section>
+        </section>
+        <div className="container">
+          <section className="near-places places">
+            <h2 className="near-places__title">Other places in the neighbourhood</h2>
+            <OffersList offers={[...nearOffers]} />
+          </section>
+        </div>
+      </main>
     );
   } else {
     return (
