+++ project/src/store/index.ts
@@ -1,14 +1,20 @@
-import {reducer} from './reducer';
 import {configureStore} from '@reduxjs/toolkit';
 import {AppDispatch, Store} from '../types/store';
 import {TypedUseSelectorHook, useDispatch, useSelector} from 'react-redux';
 import {createAPI} from '../services/api';
+import {userReducer} from './user/reducer';
+import {offersReducer} from './offers/reducer';
+import {offerReducer} from './offer/reducer';
 
 
 export const api = createAPI();
 
 export const store = configureStore({
-  reducer,
+  reducer: {
+    user: userReducer,
+    offers: offersReducer,
+    offer: offerReducer,
+  },
   middleware: (getDefaultMiddleware) =>
     getDefaultMiddleware({
       thunk: {
